requirements.txt
fastapi
uvicorn
python-dotenv
supabase-py
requests
httpx
vllm
